# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Restaurant Ordering Platform API",
    "version": "1.0.0",
    "description": "Every response is `{ success: true, data }` or `{ success: false, error: { code, message } }`. All money fields are integer kobo. Admin routes require the admin_session cookie (log in via POST /api/admin/auth/login below, then Swagger UI will carry the cookie automatically on same-origin requests). State-changing admin requests also require a matching Origin header, which the browser sets automatically when this page is served from the same app. Checkout is accounts-only: customer routes require the customer_session cookie (obtained via POST /api/auth/otp/request then POST /api/auth/otp/verify below), separate from admin_session and carried the same way by Swagger UI on same-origin requests."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this docs page"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      },
      "customerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "customer_session"
      },
      "reconcileSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "RECONCILE_SECRET, for the internal reconcile cron only."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/paystack/webhook": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Paystack webhook receiver",
        "description": "Signature-verified via the x-paystack-signature header (HMAC-SHA512 of the raw body). Not callable meaningfully from Swagger UI — Paystack signs the exact raw bytes it sends, so a hand-built request here will always fail signature verification. Included for completeness.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acknowledged (event processed, ignored, or a safe no-op on an unknown reference)."
          },
          "401": {
            "description": "Invalid or missing signature."
          },
          "500": {
            "description": "Processing failed — Paystack will retry."
          }
        }
      }
    },
...
    "/api/internal/reconcile-payments": {
      "post": {
        "tags": [
          "Internal"
        ],
        "summary": "Run the payment reconciler + table sweep",
        "description": "Not part of the public/customer/admin surface. Authenticated by `Authorization: Bearer $RECONCILE_SECRET`. Re-verifies stale PENDING payments against Paystack and applies missed outcomes, heals paid-but-unadvanced orders, escalates the genuinely-stuck to PaymentIssue rows, and sweeps expired RateLimit/OTP rows. Called every ~10 min by a Netlify scheduled function; safe to invoke by hand. GET behaves identically.",
        "security": [
          {
            "reconcileSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Run summary (counts).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or wrong bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}